Consolidate UI components into `ui/` folder and update documentation.
- Moved `ui/widgets/memory_watch.py` to `ui/memory_watch.py` to flatten the directory structure.
- Removed `ui/widgets/` directory.
- Updated `ui/main_window.py` to import from the new location.
- Updated `USER_GUIDE.md` to reflect the current state of the application (mock data, local logs).
- Updated developer documentation (`implementation_roadmap.md`, `ui_planning.md`) to reflect the new structure.
